Move drawPineapple to shapes/pineapple
